split head and home poge in components
